requirements.txt
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gspread==6.2.1
h11==0.16.0
httplib2==0.31.0
idna==3.11
imap-tools==1.11.0
oauth2client==4.1.3
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.3.1
pypdfium2==5.2.0
pyperclip==1.11.0
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
python-anticaptcha==1.0.0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
